backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==11.3.0
requests==2.32.4
gunicorn==21.2.0
whitenoise==6.6.0